requirements.txt
scikit-learn>=0.19.1
numpy>=1.14.2
matplotlib>=2.2.2
pyyaml>=3.12
Keras>=2.1.5
tensorflow>=1.7.0